VectorBuilder.Builder

Documentation

data Builder element

An abstraction over the size of a vector for the process of its construction.

It postpones the actual construction of a vector until the execution of the builder.

empty :: Builder element

Empty builder.

singleton :: element -> Builder element

Builder of a single element.

size :: Builder element -> Int

Gets the size of a Builder.

vector :: Vector vector element => vector element -> Builder element

Builder from an immutable vector of elements.

Supports all kinds of vectors: boxed, unboxed, primitive, storable.

foldable :: Foldable foldable => foldable element -> Builder element